archive/requirements.txt
# OpenAI Agent SDK
openai>=1.50.0
agents>=0.1.0

# Data handling and validation
pydantic>=2.0.0
pandas>=2.0.0

# File processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0

# UI Framework
streamlit>=1.30.0

# Configuration management
python-dotenv>=1.0.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# Logging
loguru>=0.7.0

# Date/Time handling
python-dateutil>=2.8.0
